refactor: replace SortButton with SortSelector for improved sorting functionality
